+++ include/nighthawk/common/BUILD
@@ -64,20 +64,6 @@ envoy_basic_cc_library(
     ],
 )
 
-envoy_basic_cc_library(
-    name = "request_source_plugin_config_factory_lib",
-    hdrs = [
-        "request_source_plugin_config_factory.h",
-    ],
-    include_prefix = "nighthawk/common",
-    deps = [
-        ":request_source_lib",
-        "//api/request_source:request_source_plugin_cc_proto",
-        "@envoy//include/envoy/common:base_includes",
-        "@envoy//include/envoy/config:typed_config_interface",
-        "@envoy//source/common/api:api_lib_with_external_headers",
-    ],
-)
 
 envoy_basic_cc_library(
     name = "request_source_lib",
